+++ IntelliHome-Backend/Data/Models/VEU/VehicleCharger.cs
@@ -9,7 +9,7 @@ namespace Data.Models.VEU
 {
     public class VehicleCharger : SmartDevice
     {
-        public Decimal Power { get; set; }
+        public Double Power { get; set; }
         List<VehicleChargingPoint> ChargingPoints { get; set; }
         public Int32 FreeChargingPoints => ChargingPoints.Count(x => x.IsFree == true);
